requirements.txt
appdirs==1.4.4
apturl==0.5.2
beautifulsoup4==4.10.0
bidict==0.21.3
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
evdev==1.3.0
google==3.0.0
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-core==2.0.0
google-cloud-datastore==2.1.6
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
gspread==4.0.1
html5lib==1.0.1
httplib2==0.19.1
idna==2.8
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libcst==0.3.20
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
matplotlib==3.4.3
MouseInfo==0.1.3
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
ncbi-genome-download==0.3.0
netifaces==0.10.4
numpy==1.21.2
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
pandas==1.3.3
Pillow==7.0.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGetWindow==0.0.9
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.14.1
PyRect==0.1.4
pyRFC3339==1.1
PyScreeze==0.1.27
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
python-engineio==4.2.1
python-magic==0.4.16
python-socketio==5.4.0
python3-xlib==0.15
PyTweening==1.0.3
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scipy==1.7.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==2.2.1
systemd-python==234
tqdm==4.62.3
typing-extensions==3.10.0.2
typing-inspect==0.7.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
wadllib==1.3.3
webencodings==0.5.1
websocket-client==1.2.1
xkit==0.0.0
